Use datalad's long opts rather than short opts
IMO, in a tutorial, it's better to be overly explicit. And it's more self
documenting this way. Users who don't like typing will quickly discover the
short opts.

I only touched DataLad's options and not any of the other tools invoked. I also
intentionally left '-m' unexpanded because of the context and the intro example
was sufficiently clear.