Large changes to "Working with containers"
Not an overhaul, but this section required a lot more changes than the others,
and a lot more rethinking. In addition to the usual cleanup:

* I moved the "everything is scary" paragraph from the previous section into the
  current one, and reworked it so the reader understood /why/ we were scaring
  them.
* Change the approach to the "cp that file" thing

I'm not convinced that the title of this section is correct, as a container is
only used once, and it's more about `datalad run` and `datalad container-run`.