Cleanup "A reproducible GLM demo analysis"
In addition to the normal cleanup, this was largely a fix of tone. It is not
nice to string users along, less so when you joke about it. Stop promising one
more small step, stop pretending we're exasperated and at the last pre-step, one
to have one more. Be dry, be to the point, they will provide their own
narrative - exasperated or not.