reproin: Don't hardcode HEAD~10 in afterlife solution
Using HEAD~10 is fragile.  It needs to be updated if any of the steps
above are modified, and it assumes the solutions above were achieved
in exactly the same way, but it would be possible to get to the same
state with a different number of commits (e.g., by using 'datalad
download-url --path code/').